+++ model/schema_order_test.go
@@ -5,9 +5,11 @@ import "testing"
 func TestGetOrder(t *testing.T) {
 	// TODO: Schema.GetAscRows,GetAscRowsAt
 	// TODO: Schema.GetDescRows, GetDescRowsAt
+	// TODO: Schema.GetOrderedRows, GetOrderedRowsAt
 }
 
 func TestTopOrder(t *testing.T) {
 	// TODO: Schema.TopAscRows, TopAscRowsAt
 	// TODO: Schema.TopDescRows, TopDescRowsAt
+	// TODO: Schema.TopOrderedRows, Schema.TopOrderedRowsAt
 }