+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## v0.3.31 - Multi-Column Order
+  * Commit: 2026-05-18 09:56
+  * Schema GetOrderedRows
+  * Schema TopOrderedRows
 ## v0.3.30 - Secrets Package 
   * Commit: 2026-05-16 23:38
   * Secrets package